Renaming changes in player state
- Put definitions in player_state.h into their own player_state
  namespace
- Drop the Player* prefix for all player state related structs
- Rename opponent_* to enemy_* in player state
- Create player state's own soldier state enum
- Drop the STATE_EXPORT macro in player state as no structs there have
  methods

The above changes ensure that the player state is no longer dependent
on any other headers in the state module. It also makes it easier to
use for the players as the Player* prefix has been dropped and enemy_*
is consistently used everywhere.

Update all usages and tests of the player state structs